helpers: cleanup test files
