+++ packages/helpers/tests/chunked-zkey.test.ts
@@ -1,6 +1,8 @@
+import fs from 'fs';
+import path from 'path';
 import { StringDecoder } from "string_decoder";
 import _localforage from "localforage";
-import { downloadFromFilename, downloadProofFiles } from "../src/zkp";
+import { downloadFromFilename, downloadProofFiles, uncompressGz as uncompress } from "../src/chunked-zkey.js";
 import { server } from './mocks/server.js'
 import { MOCK_BASE_URL } from "./mocks/handlers.js";
 
@@ -26,6 +28,28 @@ const decodeArrayBufferToString = (buffer: ArrayBuffer): string =>  {
   const str = decoder.write(Buffer.from(buffer));
   return str;
 }
+
+const getCompressedTestFile = (): ArrayBuffer => {
+  const buffer = fs.readFileSync(path.join(__dirname, `test-data/compressed-files/compressed.txt.gz`));
+  return buffer;
+}
+
+const getUncompressedTestFile = (): ArrayBuffer => {
+  const buffer = fs.readFileSync(path.join(__dirname, `test-data/compressed-files/uncompressed-value.txt`));
+  return buffer;
+}
+
+describe('Uncompress GZ file', () => {
+  test('Uncompresss a GZ file', async () => {
+    const decoder = new StringDecoder('utf8');
+    const compressedArrayBuffer: ArrayBuffer = getCompressedTestFile();
+    const expectedArrayBuffer: ArrayBuffer = getUncompressedTestFile();
+    const expectedString = decoder.write(Buffer.from(expectedArrayBuffer));
+    const uncompressedArrayBuffer = await uncompress(compressedArrayBuffer);
+    const uncompressedString = decoder.write(Buffer.from(uncompressedArrayBuffer));
+    expect(uncompressedString).toBe(expectedString);
+  });
+});
 
 describe('Test zkp fetch and store', () => {
